Add README and use folders
